feat: add companies air
