+++ apps/naugrim/src/components/Video/index.tsx
@@ -15,8 +15,8 @@ export const Video = () => {
     <section className="relative z-10 py-16 md:py-20 lg:py-28">
       <div className="container">
         <SectionTitle
-          title="We are ready to help"
-          paragraph="There are many variations of passages of Lorem Ipsum available but the majority have suffered alteration in some form."
+          title="Sempre ao Seu Lado"
+          paragraph="Sua tranquilidade é a nossa missão, em qualquer lugar, a qualquer hora"
           center
           mb="80px"
         />
