+++ apps/naugrim/src/data/types/Features.ts
@@ -1,0 +1,12 @@
+export interface Feature {
+  id: number
+  icon: JSX.Element
+  title: string
+  content: string
+}
+
+export interface FeaturesProps {
+  features: Feature[]
+  title?: string
+  subTitle?: string
+}
